Add files via upload
Hello! My code includes the step by step process for determining whether paired integer numbers are equivalent. This considers their placement in the ones, tens, and hundreds place. The first lines of coding are constructed in order to get two input numbers. The next section is coding that corresponds the equivalency of the two integer numbers. To complete the program, the if and else statements are utilized in support of the boolean expressions which will output the result of the paired integer numbers.