Add files via upload
Hello! My code is a step by step process that determines the equivalence of two integer numbers. The first lines of coding focuses on gathering the two input numbers. The next section of coding corresponds the numbers together by the ones, tens, and hundreds place. Before running the code, the if and else statements are utilized in support of the Boolean expression which provides an output to the integer numbers that are paired.