Update Lab1Part1.csproj
Hello! My code is a step by step process that determines the equivalence of two integer numbers. The first few lines of coding serves the purpose of getting two input numbers. The second section of coding corresponds the integers by the ones, tens, and hundredths place. Before running the code, the if and else statements support the Boolean expression which provides the output as to whether the paired integers are equivalent or otherwise.